backend/requirements-docker.txt
fastapi==0.135.1
uvicorn==0.41.0
langchain==1.2.12
langchain-groq==1.1.2
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-core==1.2.18
langchain-text-splitters==1.1.1
chromadb==1.5.5
pypdf==6.8.0
python-multipart==0.0.22
python-dotenv==1.2.2
sentence-transformers==5.2.3
torch==2.10.0+cpu